统一类型
