--- include/elastic/serialize/types.hpp
@@ -1,19 +1,0 @@
-#pragma once
-#include <optional>
-#include <string>
-#include <vector>
-
-namespace elastic
-{
-	using int8 = int8_t;
-
-	using int16 = int16_t;
-
-	using int32 = int32_t;
-
-	using int64 = int64_t;
-
-	using string = std::string;
-
-	using bytes = std::vector<std::byte>;
-} // namespace elastic
